Move DropTarget to its own module
