+++ src/win/mod.rs
@@ -1,3 +1,4 @@
+mod drop_target;
 mod keyboard;
 mod window;
